requirements.txt
pygerduty==0.32.1
PyYAML==3.11
requests==2.9.1
slacker==0.9.0
six==1.9.0
urllib3==1.13.1
